sdlc_langchain/requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-community>=0.0.10

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0

# Async and Parallel Processing
aiohttp>=3.9.0
aiofiles>=23.0.0

# Memory and Storage
redis>=5.0.0

# Data Handling
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0

# UI and Output
rich>=13.0.0
tqdm>=4.66.0

# Utilities
tenacity>=8.2.0
cachetools>=5.3.0
